requirements.txt
numpy
sounddevice
soundfile
faster-whisper
gTTS
pydub
requests
python-dotenv
loguru
pytest
hf_xet
